requirements.txt
Pillow==7.1.1
PyQt5==5.14.2
PyQt5-sip==12.7.2
pyzbar==0.1.8
